Usage: remove support of option file
Being able to override settings using an options file was a confusing
feature that provided little benefit. There was no validation performed
on it's contents and it widened the scope of misusing of the tool.

Users are expected to use the config as the defacto source of
configuration. Examples will be updated in due course.