Usage: remove support of external namespace
Supporting <external dir>/<name>/ as a location for sub-components has
questionable benefit and adds unnecessary variables and processing to
various stages. If this flexibility is needed in the future, it can be
added. Supporting <external dir> as an additional root dir of
sub-components provides an acceptable location for custom changes
outside of in-tree / upstream.