Add python engine and classes for core support
Delegate config file, metadata and layer management to a bespoke python
engine which is invoked by a top level helper. The helper is the entry
point that scripts and users can use to develop, inspect and interact
with config files and the layer sub-system.

The engine provides the means for files to define, describe, set and
check variables using X-Env- field metadata primitives. The metadata
(based on DEB822) can be be embedded in files using defined commented
syntax or native, and allows layers to declare relationships between
them and perform validation of variables they declare or which are
required from other sources. This enables dependencies between layers
to be supported, as well as allowing categorisation, grouping, validation
and vastly improving documentation.

The format of the metadata lends itself well to being able to
generate documentation for files that contain it. It is envisaged that
future commits will leverage this to build a documentation tree for all
supported layers and their variables.