layer: Clarify apt cache dir wording
There are a few different ways that a reusable dir for apt to use
between builds can be achieved. The simplest way is to bind mount a
directory into the chroot and make sure apt doesn't clear the cache
upon build completion, but this introduces some scripting complexity.
There are some additional apt options that also need to be passed,
too. bdebstrap < 0.6.0 does not support passing mmdebstrap's skip
options in YAML which would be needed to skip wiping the cache.
Local testing with additional scripts and the following options has
proven that a cache dir can be bind mounted at setup and unmounted
at cleanup before mmdebstrap runs its cache cleanup, therefore allowing
the contents to be reused between builds.

--aptopt Debug::pkgAcquire true
--aptopt Debug::Acquire::file true

Using this feature is currently not advised, so explicitly state it's
experimental.