core: device and image agnostic builds pt1
Significant updates across the board to remove various assumptions that
a build will always contain device and/or image support. This is not the
case: rpi-image-gen needs to be able to accomodate builds which may
required to, for example, generate a device agnostic filesystem tarball
suitable for containerised deployments, or to create a directory ready
for use as a build chroot.

To support this, add new base layers to provide logically separated
variables. Simplify hooks by using variables to hold their paths.

Relocate sbom and deploy hooks into the layer lib, and use placeholder
variables to allow them to be located more easily by the core.

Improve deployment functionality by installing more files and creating
a simple json manifest.