core: device and image agnostic builds pt2
Complete the support of device and image agnostic builds, therefore
enabling rpi-image-gen to create simple things like a filesystem
tarball which can be utilised by container utilities.

Add a new example which does just that, creating a 'build chroot'
tarball and importing it as a container, then spinning up the container
to compile a src pkg with GNU autotools.