layer: Add more device layers
Add layers for other rpi devices. All except cm5 simply pull in the
generic family64 layer which will install kernel and boot firmware
compatible with any ARMv8 device. The cm5 layer is essentially a
duplicate of rpi5 (2712 kernel). The reason for having 'family'
devices split like this is so a user can have completely
device-specific items in it, which may not be applicable to their
deployment.