engine: Add support for dynamic dependencies pt2
Because documentation generation follows the same code path as
build-time when it comes to instantiating the layer manager, if a layer
refers to an env var when specifying dependencies, unless that var is
set when the documentation is generated, the layer will be skipped
because it can't be found. This is the desired behaviour for build-time
but not for docgen-time.

To combat this, pass a flag down from the top level so we can be more
lenient when generating documentation for layers with dynamic
dependencies.

Also, update the layer jinja template to strip whitespace. Next time we
autogenerate docs there will be removal of blank lines in some cases,
but that's better anyway.