test: Add support for linting layers with dynamic deps
Layers with dynamic dependencies will fail the bulk lint all test
unless the appropriate variable(s) are set in the env when the test
runs. Add support for that, so a layer file can be linted with a
defined env by sourcing a file matching the basename of the layer
file. Files to seed the env can be put in test/layer/env/dist/ and must
have a .env suffix.