AB(emmc): write-protect the partition table (properly)
The recent major surgery removed the partition table protection feature
for AB on eMMC so add support back in, declaring the variable in the
image layer and defaulting it to y if the image is built for emmc
storage.

Also broken (separately) was the usage of the AB slot helper script
which was updated to emit more information about the boot device.
This additional information wasn't being catered for in the emmc
protection script which resulted in the script always exiting, meaning
that no protection was ever actually applied.

Lastly, the AB genimage template was placing the GPT Partition Entry
Array at offset +8M in addition to specifying an 8MB alignment. So even
if the first 8MB were protected, the protection only covered:

- The Protective MBR (LBA 0)
- The Primary GPT header (LBA 1)

To actually protect the GPT, the Primary Partition Entry Array should be
inside the protected region. Using the default sectors 2–33 is fine for
this. The alignment of 8MB is still applied so partitions are correctly
created on 8MB boundaries.

Verified by booting the AB image, verifying the protection with
mmc-utils (mmc writeprotect user get /dev/mmcblk0), trying to blow away
the partition table (wipefs -a -f /dev/mmcblk0) and rebooting. Reboot
is successful.