engine: use doc mode when listing layers
When not in doc mode, layers are automatically ignored if they cannot
be located, which is a problem for layers that specify dynamic
dependencies when the env vars they reference aren't present in the env.
We need all valid layers to be listed on the cmdline, so use doc mode
specifically when doing that. Layers still get linted.