AB: Dynamic slot selection now use GPT labels with fallback
Refactor the AB image layer to use GPT labels for slot identification
and run-time selection, with a fallback to a static mapping for devices
or configurations that may not or cannot use GPT. Add companion doc.

Move all the run-time resources responsible for managing the AB slots
to a new layer. This includes udev rules, slot helper scripts and
initramfs-tools fragments. Tighten and improve the existing udev rules
that trigger slot block device creation, ensuring that systems that
use DM and non-DM devices can be supported with and without GPT labels.

Rework the pmap helper to generate the static slot map from the PMAP.
Remove UUID retrieval and instead, parse the JSON to output slot map
triplets compatible with the static map udev helper.

Only generate one pair of slot filesystem images. This ensures that
both slots are bit for bit identical. This has several benefits and
does not impact users of stable block device identifiers.

As IDP does not yet honour GPT labels in the image, install a static map
so that images provisioned with IDP still boot. This can be removed once
the IDP host side is updated.

We still need the rpi-fw-crypto plumbing before we can support full crypt,
so restrict IDP configuration to clear and hybrid.