engine: Add tagged path support
Extend the layer mgmt class to accept tags in the search path, add a
new relative path output format which includes those tags, and use it
in the top level script. By supporting tags in the layer search path it
means a client can map a layer path from layer discovery (bootstrap) to
build, meaning that it's possible to translate those paths into a
container environment.