layer: Add generator and a reproducible fs example
envcp is a simple dynamic layer generator that substitutes placeholder
tags in the input template with environment variables contents or
defaults. It can be extended as required, and currently only serves as
a generator used to contruct Debian apt snapshot urls.

Example reproducible_debfs demonstrates how to use a dynamic layer and
the envcp generator so that snapshotted Debian binaries can be used to
create the base system, eg for a container image.

Running this build at any time will yield identical binaries installed
in the filesystem.